add createBookingDurations function
